+++ yapconf/sources.py
@@ -1,12 +1,24 @@
 # -*- coding: utf-8 -*-
 import abc
+import copy
+import hashlib
 import json
 import os
+import threading
+import time
 
 import six
+from watchdog.observers import Observer
 
 import yapconf
 from yapconf.exceptions import YapconfLoadError, YapconfSourceError
+from yapconf.handlers import FileHandler
+
+if yapconf.kubernetes_support:
+    from kubernetes.watch import watch
+
+if yapconf.etcd_support:
+    from etcd import EtcdWatchTimedOut
 
 
 def get_source(label, source_type, **kwargs):
@@ -103,6 +115,7 @@ class ConfigSource(object):
     def __init__(self, label):
         self.label = label
         self.validate()
+        self._continue = True
 
     @abc.abstractmethod
     def validate(self):
@@ -112,7 +125,7 @@ class ConfigSource(object):
     def generate_override(self, separator='.'):
         """Generate an override.
 
-        Uses ``_get_data`` which is expected to be implemented by each child
+        Uses ``get_data`` which is expected to be implemented by each child
         class.
 
         Returns:
@@ -122,7 +135,7 @@ class ConfigSource(object):
             YapconfLoadError: If a known error occurs.
 
         """
-        data = self._get_data()
+        data = self.get_data()
         if not isinstance(data, dict):
             raise YapconfLoadError(
                 'Invalid source (%s). The data was loaded successfully, but '
@@ -131,8 +144,62 @@ class ConfigSource(object):
             )
         return self.label, yapconf.flatten(data, separator=separator)
 
+    def _spawn_and_watch(self, handler):
+        thread = self._spawn_watcher(handler)
+        while self._continue:
+            if thread.isAlive():
+                time.sleep(5)
+            else:
+                thread = self._spawn_watcher(handler)
+
+    def _spawn_watcher(self, handler):
+        thread = threading.Thread(
+            name=self.label + '-watcher',
+            target=self._watch,
+            args=(handler, self.get_data()),
+        )
+        thread.setDaemon(True)
+        return thread
+
+    def watch(self, handler, eternal=False):
+        """Watch a source for changes. When changes occur, call the handler.
+
+        By default, watches a dictionary that is in memory.
+
+        Args:
+            handler: Must respond to handle_config_change
+            eternal: Spawn eternal watch, or just a single watch.
+
+        Returns:
+            The daemon thread that was spawned.
+
+        """
+        if eternal:
+            thread = threading.Thread(
+                name=self.label + '-watcher-eternal',
+                target=self._spawn_and_watch,
+                args=(handler,)
+            )
+            thread.setDaemon(True)
+        else:
+            thread = self._spawn_watcher(handler)
+        thread.start()
+        return thread
+
+    def _hash_data(self, data):
+        return hashlib.md5(repr(data).encode('utf-8')).hexdigest()
+
+    def _watch(self, handler, data):
+        current_hash = self._hash_data(data)
+        while self._continue:
+            new_hash = self._hash_data(self.get_data())
+            if current_hash != new_hash:
+                handler.handle_config_change(self.data)
+                current_hash = new_hash
+            time.sleep(1)
+
     @abc.abstractmethod
-    def _get_data(self):
+    def get_data(self):
         pass
 
 
@@ -155,7 +222,7 @@ class DictConfigSource(ConfigSource):
                 'dictionary. Got: %s' % (self.label, self.data)
             )
 
-    def _get_data(self):
+    def get_data(self):
         return self.data
 
 
@@ -187,7 +254,24 @@ class JsonConfigSource(ConfigSource):
                 'JSON config source.' % self.label
             )
 
-    def _get_data(self):
+    def _watch(self, handler, data):
+        if self.filename:
+            file_handler = FileHandler(
+                filename=self.filename,
+                handler=handler,
+                file_type='json'
+            )
+            observer = Observer()
+            directory = os.path.dirname(self.filename)
+            observer.schedule(file_handler, directory, recursive=False)
+            observer.start()
+            observer.join()
+        else:
+            raise YapconfSourceError(
+                'Cannot watch a string json source. Strings are immutable.'
+            )
+
+    def get_data(self):
         if self.data is not None:
             return json.loads(self.data, **self._load_kwargs)
 
@@ -222,7 +306,19 @@ class YamlConfigSource(ConfigSource):
                 'config source.' % self.label
             )
 
-    def _get_data(self):
+    def _watch(self, handler, data):
+        file_handler = FileHandler(
+            filename=self.filename,
+            handler=handler,
+            file_type='yaml'
+        )
+        observer = Observer()
+        directory = os.path.dirname(self.filename)
+        observer.schedule(file_handler, directory, recursive=False)
+        observer.start()
+        observer.join()
+
+    def get_data(self):
         return yapconf.load_file(
             self.filename,
             file_type='yaml',
@@ -237,6 +333,9 @@ class EnvironmentConfigSource(DictConfigSource):
     def __init__(self, label):
         super(EnvironmentConfigSource, self).__init__(label, os.environ.copy())
         self.type = 'environment'
+
+    def get_data(self):
+        return copy.deepcopy(os.environ)
 
 
 class EtcdConfigSource(ConfigSource):
@@ -265,7 +364,7 @@ class EtcdConfigSource(ConfigSource):
                  type(self.client))
             )
 
-    def _get_data(self):
+    def get_data(self):
         result = self.client.read(key=self.key, recursive=True)
         if not result.dir:
             raise YapconfLoadError(
@@ -274,11 +373,19 @@ class EtcdConfigSource(ConfigSource):
             )
 
         data = {}
-        for child in result:
+        for child in result.children:
             keys = self._extract_keys(child.key)
             self._add_value(data, keys, child.value)
 
         return data
+
+    def _watch(self, handler, data):
+        while self._continue:
+            try:
+                self.client.read(key=self.key, wait=True, recursive=True)
+                handler.handle_config_change(self.get_data())
+            except EtcdWatchTimedOut:
+                pass
 
     def _add_value(self, data, keys, value):
         for i, key in enumerate(keys):
@@ -342,7 +449,7 @@ class KubernetesConfigSource(ConfigSource):
                 (self.name, yapconf.FILE_TYPES)
             )
 
-    def _get_data(self):
+    def get_data(self):
         result = self.client.read_namespaced_config_map(self.name,
                                                         self.namespace)
         if self.key is None:
@@ -362,3 +469,22 @@ class KubernetesConfigSource(ConfigSource):
         else:
             raise NotImplementedError('Cannot load config with type %s' %
                                       self.config_type)
+
+    def _watch(self, handler, data):
+        w = watch.Watch()
+        for event in w.stream(
+            self.client.list_namespaced_config_map,
+            namespace=self.namespace
+        ):
+            config_map = event['object']
+            if config_map.metadata.name != self.name:
+                continue
+
+            if event['type'] == 'DELETED':
+                raise YapconfSourceError(
+                    'Kubernetes config watcher died. The configmap %s was '
+                    'deleted.' % self.name
+                )
+
+            if event['type'] == 'MODIFIED':
+                handler.handle_config_change(self.get_data())