+++ yapconf/spec.py
@@ -9,8 +9,9 @@ import yapconf
 from box import Box
 from yapconf.docs import generate_markdown_doc
 from yapconf.exceptions import (YapconfItemNotFound, YapconfLoadError,
-                                YapconfSpecError)
-from yapconf.items import YapconfDictItem, from_specification
+                                YapconfSourceError, YapconfSpecError)
+from yapconf.handlers import ConfigChangeHandler
+from yapconf.items import YapconfDictItem, YapconfListItem, from_specification
 from yapconf.sources import get_source
 
 if sys.version_info.major < 3:
@@ -186,6 +187,27 @@ class YapconfSpec(object):
         """
         self._sources[label] = get_source(label, source_type, **kwargs)
 
+    def find_item(self, fq_name):
+        """Find an item in the specification by fully qualified name.
+
+        Args:
+            fq_name (str): Fully-qualified name of the item.
+
+        Returns:
+            The item if it is in the specification. None otherwise
+
+        """
+        names = fq_name.split(self._separator)
+        current = self._yapconf_items
+        for name in names:
+            if isinstance(current, (YapconfDictItem, YapconfListItem)):
+                current = current.children
+
+            if name not in current:
+                return None
+            current = current[name]
+        return current
+
     def get_item(self, name, bootstrap=False):
         """Get a particular item in the specification.
 
@@ -303,6 +325,36 @@ class YapconfSpec(object):
         overrides = self._generate_overrides(*args)
         config = self._generate_config_from_overrides(overrides, bootstrap)
         return Box(config)
+
+    def spawn_watcher(self, label, target=None, eternal=False):
+        """Spawns a config watcher in a separate daemon thread.
+
+        If a particular config value changes, and the item has a
+        ``watch_target`` defined, then that method will be called.
+
+        If a ``target`` is passed in, then it will call the ``target``
+        anytime the config changes.
+
+        Args:
+            label (str): Should match a label added through ``add_source``
+            target (func): Should be a function that takes two arguments,
+            the old configuration and the new configuration.
+            eternal (bool): Determines if watcher threads should be restarted
+            if they die.
+
+        Returns:
+            The thread that was spawned.
+
+        """
+
+        if label not in self._sources:
+            raise YapconfSourceError(
+                'Cannot watch %s no source named %s' % (label, label)
+            )
+
+        current_config = self._sources[label].get_data()
+        handler = ConfigChangeHandler(current_config, self, target)
+        return self._sources[label].watch(handler, eternal)
 
     def migrate_config_file(
         self,